service: add Trigger(name) to trigger bundle rebuild
Adds the machinery to pool and service to trigger bundle rebuilds
by name. If the bundle worker is queued, it'll be pulled into the
front of the line, to be picked up next; if it's currently not in
the queue, it must be executing: if so, we're overriding the next
deadline to be _now_, causing an immediate re-run of the build
process.

For now, this is only wired up with signal handling: send a HUP signal
to rebuild. This is because we haven't discussed proper permissions
and HTTP API design for this yet.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>