builder: support "mounts", i.e. rewrites of rego and data in sources
This change allows users to drive certain bundle data and policy rewrites
from the configuration.

We're calling it "mounts", and they come with (optional) sub and prefix
settings. A "sub" will select a subtree of `data` and put it into a new
"prefix". E.g. the following mount,

    sub: data.library
    prefix: data.imported.lib.v1

would change all `package` statements of the rego files within this
source and sources it depends on. A package like "data.library.authz"
would be rewritten to "data.imported.lib.v1.authz". All references are
adjusted.

Mounts are aggrated, and applied in order, if the source that you depend
on has a mount configuration, too.

Subtrees of data that are not selected by a "sub" are carried along.
It is possible to prefix everything that comes with a source by using
sub "data" (and prefix "data.source1" for example).

Fixes #30.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>