database, tokens: env expansions (#144)
* config: nitpicks in secret.go

* database: expand tokens (env vars) before loading

This feels somewhat preliminary until we introduce a proper
tokens CRUD API. Right now, the database tables for tokens are
filled when the config file is ingested.

* database: expand env vars in database configs

I hope I've covered all of them. This is slightly different than
tokens, as the DB config isn't persisted to the database...

* database: expand env vars in datasource config _files_

This only applies to datasources when configured in a file: they expansion
happens when the datasource definition is carried over into the database.


Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>