builder: support "mounts", i.e. rewrites of rego and data in sources  (#135)
This change allows users to drive certain bundle data and policy rewrites
from the configuration.

We're calling it "mounts", and they come with (optional) sub and prefix
settings. A "path" will select a subtree of `data` and put it into a new
"prefix". E.g. the following mount,

```yaml
path: data.library # default "data", noop
prefix: data.imported.lib.v1 # default "data", noop
```

would change all `package` statements of the rego files within this
source and sources it depends on. A package like "data.library.authz"
would be rewritten to "data.imported.lib.v1.authz". All references are
adjusted. The "data." prefix can be omitted, so the above is equivalent
to

```yaml
path: library
prefix: imported.lib.v1
```

Mounts are aggregated, and applied in order, if the source that you depend
on has a mount configuration, too.

Subtrees of data that are not selected by a "path" are ignored.
It is possible to prefix everything that comes with a source by using
path "data" (or no path at all) and prefix "data.source1" for example.

Note that the substitution for data depends on the file system at the
moment. If you have a data.yaml like this in a/b/data.yaml,

```yaml
c:
  d: foo
```

then the `path` `data.a.b.c` isn't going to work; only path `data.a` or
`data.b` would be supported. This is a restriction we might revisit
later.

Fixes https://github.com/open-policy-agent/opa-control-plane/issues/30.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>