bundles: make "rebuild interval" configurable
This is the time spent doing nothing between successful bundle builds. If a
build fails, it'll be retried in 30s (like before).

The introduction of the config field was trickier than expected: it's added
as a string in YAML/JSON files, and the API requests, it's stored as a string
in the database. But when used, it's parsed as a time.Duration.

I'ved opted for storing the string in the database (not an int) because it
might be slightly nicer to return what the user has sent, without messing it
up. If you configure your bundle to have a `rebuild_interval` of `90m`, then
that's what we'll report. It won't magically turn into `1h30m` or a large
integer representation of seconds.

All the UnmarshalJSON/UnmarshalYAML business is to make reading API requests
and config files properly. PrepareJSONSchema is to make the reflection-based
JSON schema machinery accept a string where the config field is
`config.Duration`.

Fixes #102.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>