authz: principals need to be per-tenant (since they have roles)
All the resources that are affected by, say, Bob having role "viewer", are
in tenants. It doesn't make much sense to have Bob be a "viewer" of all
tenants. This now behaves like resource_permissions: If bob creates a thing,
they're only the owner of that (named) thing in the specific tenant it was
created in.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>