introduce tenants in the database layer (#202)
This is preliminary work enabling a multi-tenant future for OCP. What happens in
this change:

1. All the entity tables that used a "name" as ID now use their integer ID as
   primary key, including in relation tables (bundles requiring sources, etc)
2. Entity names are now unique within a "tenant". Currently, only the "default"
   tenant exists, and tenant-related information is not exposed through the
   HTTP API.
3. Almost all tables are recreated in the migration. This is done because it's
   tricky to update cross-table constraints in a portable way across
   sqlite/mysql/postgres.

Late addition: Principals are tenant-scoped, too, now:

All the resources that are affected by, say, Bob having role "viewer", are
in tenants. It doesn't make much sense to have Bob be a "viewer" of all
tenants. This now behaves like resource_permissions: If bob creates a thing,
they're only the owner of that (named) thing in the specific tenant it was
created in.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>