migrations: name all constraints
This has no effect on the existing data -- we don't rename the constraints in
migrations (that's kind of the point, we cannot do this easily). BUT since v0.3.0
will recreate all tables, we'll use this opportunity to introduce names.

This change adds `tokens` to the list of re-created tables, so that we have full
coverage of named constraints, as asserted by the new test.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>